Move SCM binaries under the Binaries namespace
