+++ src/Rocketeer/Traits/BashModules/Binaries.php
@@ -19,6 +19,9 @@ namespace Rocketeer\Traits\BashModules;
  * @method \Rocketeer\Binaries\PackageManagers\Npm npm()
  * @method \Rocketeer\Binaries\Php php()
  * @method \Rocketeer\Binaries\Phpunit phpunit()
+ * @method \Rocketeer\Binaries\Scm\Git git()
+ * @method \Rocketeer\Binaries\Scm\Hg hg()
+ * @method \Rocketeer\Binaries\Scm\Svn svn()
  *
  * @author Maxime Fabre <ehtnam6@gmail.com>
  */
